Updated CI pipeline to execute proper SONIC test
Updated CI pipeline to fetch private submodule with ownership of the
roms to be able to execute real SONIC reinforced learning tests.

Updated the Sonic main code to be able to assert the result.

Added a unit test to verify the AI learn to plays the SONIC game with
the given reward boundaries.