Added new Python dependencies and AI logic
Replaced old OpenAI baseline and TensorFlow dependencies with newer
Stable Baselines 3 and PyTorch dependencies to be able to test and
refine reinforcement learning models for retro games.

Added a new unit test to verify that the score is decent by applying
an 'trained' agent to a Sonic the Hedgehog game environment.

Added new base ubuntu image to install proper CUDA or ROCm drivers
if the host machine has either a NVIDIA or AMD graphic card
installed.

Added a todo script to install proper PyTorch Python dependencies
depending on if the host machine has an NVIDIA or AMD graphic card
installed.

...