Created a new trained Sonic agent pipeline
Created a new train Sonic agent pipeline based on a specific Sonic
game and state aka. level and to upload the trained agent to the
artifactory containing trained agents for specific game and level.

Refined the CI pipeline to only execute Sonic the Hedgehog tests with
trained skills for a specific environment.