Refactored the normalization parameter
Refactored the normalization parameter to be true by default and by
default it sets the environment to a gray scale and with dim light
on to support dummy vectorized environment with vectorized frame
stacks by eight.

Removed the normalized environment for the verification of the
trained agent.

Changed the default policy to be cnn policy for the proximal policy
optimization algorithm.